requirements.txt
python-pptx==1.0.2
duckduckgo-search==8.0.2
groq==0.25.0
pre-commit==4.2.0
fastapi==0.115.12
starlette==0.46.2
slowapi==0.1.9
gunicorn==23.0.0
